+++ backend/src/layer/trigger/api/mod.rs
@@ -1,5 +1,6 @@
 //! This component contains the web server that enables API requests and represents the entry point for these.
 
+mod auth;
 pub mod mock;
 pub mod mutation;
 pub mod query;
